requirements.txt
python-telegram-bot==21.10
openpyxl==3.1.5
google-api-python-client==2.160.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
Pillow==11.1.0
requests==2.32.3
python-dotenv==1.0.0
